Changed from int to byte

+++ src/main/java/io/github/retrooper/packetevents/packet/PacketType.java
@@ -8,7 +8,7 @@ public class PacketType {
 
     public static class Client {
         public static class Util {
-            public static boolean isInstanceOfFlying(final int packetId) {
+            public static boolean isInstanceOfFlying(final byte packetId) {
                 return packetId == FLYING
                         || packetId == POSITION
                         || packetId == POSITION_LOOK
@@ -16,9 +16,9 @@ public class PacketType {
             }
         }
 
-        public static final Map<Class<?>, Integer> packetIds = new HashMap<Class<?>, Integer>();
+        public static final Map<Class<?>, Byte> packetIds = new HashMap<Class<?>, Byte>();
 
-        public static final int TELEPORT_ACCEPT = 0,
+        public static final byte TELEPORT_ACCEPT = 0,
                 TILE_NBT_QUERY = 1, DIFFICULTY_CHANGE = 2, CHAT = 3, CLIENT_COMMAND = 4,
                 SETTINGS = 5, TAB_COMPLETE = 6, TRANSACTION = 7, ENCHANT_ITEM = 8,
                 WINDOW_CLICK = 9, CLOSE_WINDOW = 10, CUSTOM_PAYLOAD = 11, B_EDIT = 12,
@@ -84,7 +84,7 @@ public class PacketType {
     }
 
     public static class Server {
-        public static final Map<Class<?>, Integer> packetIds = new HashMap<Class<?>, Integer>();
+        public static final Map<Class<?>, Byte> packetIds = new HashMap<Class<?>, Byte>();
         public static int SPAWN_ENTITY = 0, SPAWN_ENTITY_EXPERIENCE_ORB = 1, SPAWN_ENTITY_WEATHER = 2, SPAWN_ENTITY_LIVING = 3,
                 SPAWN_ENTITY_PAINTING = 4, SPAWN_ENTITY_SPAWN = 5, ANIMATION = 6, STATISTIC = 7,
                 BLOCK_BREAK = 8, BLOCK_BREAK_ANIMATION = 9, TILE_ENTITY_DATA = 10, BLOCK_ACTION = 11,
